road-rover-backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
asyncpg==0.24.0
sqlalchemy==1.4.22
geoalchemy2==0.10.0
alembic==1.7.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
passlib[bcrypt]==1.7.4